+++ hw/top_earlgrey/data/autogen/top_earlgrey.gen.hjson
@@ -746,7 +746,15 @@
           [
             {
               base_addr: 0x40000000
-              size_byte: 0x21000
+              size_byte: 0x1000
+            }
+            {
+              base_addr: 0x40010000
+              size_byte: 0x1000
+            }
+            {
+              base_addr: 0x40020000
+              size_byte: 0x1000
             }
             {
               base_addr: 0x40080000
